roots.rs: Automatically create a user dir in gcroots if missing
lorri symlinks gcroots into `/nix/var/nix/gcroots/per-user/$USER`. On
NixOS, that directory is automatically created for each user, but on
non-nixos it is not.

Fortunately(?), the folder is world-writable so we can create the
directory if it is missing.

Tested manually, but removing it for my user and checking whether it
is created again. Also noticed that there’s still horrible
panics (with nondescriptive IO error messages) if the directory exists
but is not writable by the current user.

Closes: https://github.com/target/lorri/issues/17