Explicitly depend on Bash, Coreutils, and their closures
stdenv uses Bash not POSIX sh for setup:

    $ nix show-derivation $(which bash)
    [...snip...]
      "builder": "/nix/store/n9acaakxahkv1q3av11l93p7rgd4xqsf-bootstrap-tools/bin/bash",
      "args": [
        "-e",
        "/nix/store/9krlzvny65gdc8s7kpb6lkx8cd02c25b-default-builder.sh"
      ]

Using POSIX sh caused syntax errors which were not caught in CI for
some reason.

Additionally, testing found that the previous attempt at including
coreutils, the run-time closure was not available and thus not usable.

This is because the Nix daemon saw the path to coreutils as a source
directory, which caused it to be copied to a new Nix store path and
then referred to.

This updated mechanism of depending on bash and coreutils is more
extensible and precise, capturing the entire closure.