socket/ReadWriter: move timeout logic to a `poll(2)` syscall
Previously, we tried to use `set_read/write_timeout` methods on
`UnixStream`s to set a timeout on the sockets. This worked fine
everywhere, except on MacOS for some elusive reason (possibly a bug in
the rust stdlib).

There is an alternative way to achieve a read/write with timeouts,
implemented in the timeout-readwrite crate:
https://docs.rs/crate/timeout-readwrite/0.2.0/source/src/writer.rs

We don’t add it as dependency, instead we specialize the interesting
`poll(2)` trick to `UnixSocket` since it’s only a few lines of code.

The timeout-readwrite library is Apache-2.0, the same license as our
project.