shell / upgrade: move to CallOpts::file via CAS
We can reify the embedded nix expression strings to disk via the `cas`
module, meaning we don’t have to pass them as expressions on the
command line anymore.

As a plus, the files get stable links which will be the same on all
machines, improving debugging (as compared to tmpdirs which are
different for each run).