project/Project: move directory creation to constructor
This has a few nice side-effects, like `Daemon` losing its lifetime
argument because we can pass a project to the `add()` method instead
of having to pass the full `Paths` object to the constructor of
`Daemon`. Also, the creation of `Roots` always succeeds now.
A whole lot of unnecessary `.unwrap()`s are gone and move to the main
function.