cargo.toml: add `atomicwrites`
Atomic file writing is required to correctly implement
`cas::ContentAddressable`.

On Linux, the crate creates a temporary directory in the current
directory (by default) and writes to that, then it hardlinks to the
real file path and deletes the temporary file. It also correctly syncs
everything to disk (files and directories).

It still uses the `tempdir` crate, which was fully merged into the
`tempfile` crate, which we also use, but that shouldn’t be a problem.