nix, builder: Fix nix output parsing
Before, the nix output was parsed into `str`/`String` before parsing
it into lines (which is the behavior of the various `.lines()`
methods).

We can, however, not assume that nix output is valid UTF-8. Take for
example the following derivation:

```
with import <nixpkgs> {};
runCommand "foo" {} ''
  dd if=/dev/urandom of=./foo bs=1M count=1
  cat ./foo
  touch $out
''
```

which prints 1 megabyte of random output to stdout (and thus to nix’s
`stderr`). This random data is not valid UTF-8, which means lorri is
going to choke on the derivation output and crash with a decoding
error.

Instead, we keep the output a `Vec<u8>`, convert it to an
`OsString` (which is a wrapper around `Vec<u8>` on POSIX systems) and
split on `\n` (which is the line separator on POSIX systems).

As a nice side-effect, we can remove the use of threads from the
`builder.rs` code (by replacing with `wait_with_output()`) and unify
the output parsing behavior).