add `lorri watch --once`
Adds a `--once` argument to lorri watch, to only run one evaluation.

`lorri install` was proposed instead, but we don’t want it to sound
like the command changes the system of the user (think `make
install`).

Ultimately, `lorri watch` is going to be replaced by `lorri daemon`
completely, then we should introduce a `lorri run` command which does
the same as `lorri watch --once`.

Co-Authored-By: Graham Christensen <graham@grahamc.com>
Co-Authored-By: Tobias Pflug <tobias.pflug@tweag.io>