CI: merge lists in travis generation instead of overwriting
This allows us to have the same phase mentioned in multiple scripts
which we later merge; the items will not get silently overwritten
anymore.

Tested first by comparing with no change (generated output was the
same) and then with having the `script` phase in multiple scripts,
which merged them as expected.