ops/upgrade: prepare for specifying arbitrary branches
This rewrites the upgrade.nix code to accept arbitrary branch
names (instead of just `master` and `rolling-release`).

It introduces a local `UpgradeSource` struct in `upgrade/mod.rs`,
which contains only the interesting information we need for calling
`upgrade.nix`, that is whether we want a local checkout or a branch.

The given command line arguments get parsed correctly instead of being
passed straight to `./upgrade.nix`, with better error messages.