ops/upgrade/upgrade.nix: don’t copy fetchedSource to nix store
We only want to read the `release.nix` file from the fetched source
repository, not import it into the store recursively; thus,
`toString`.

`release.nix` will still copy the source to the store, but it filters
src correctly, so e.g. the `.git` directory is not copied to the
store.