nix: test on 2.3.1
The older version of Nix on darwin (2.0) had a bug where it wouldn't
interpret a symlink as a `.drv`.

The upstream Travis support for newer Nixes on Darwin is a bit broken,
so this is working around that for now.

The default language changes to `minimal`, but the Linux version
overrides it to `nix`.