CI: move `cachix push` before running tests
If `cachix push` runs after the testsuite, consecutive failing tests
won’t cache their intermediate cachix results, leading to unnecessary
rebuilds when debugging the failing tests.

In order to push the dependencies of `shell.nix`, we have to pass them
out manually. As long as we don’t forget to update the dependency
list, this should be fine.