Ignore some metadata modificaiton events
We ignore metadata modificaiton events for the profiles directory
tree as it is a symlink forest that is used to keep track of
channels and nix will unconditionally update the metadata of each
link in this forest.
These ModifyKind::Metadata events are spurious and they can easily
cause a rebuild-loop when the shell.nix files  does not pin it's
version of nixpkgs or other channels.
When a Nix channel is updated we will recieve a data modification
event, so ignoring these metadata modifications will not impact
lorri's ability to correctly watch for changes.