LoopHandlerEvent
We need to communicate within the daemon process when a new listener
joins, so that we can start fanning out build events to them. That
message needs to follow the same channels as the existing
build_loop::Events do, once they get to the daemon.

The solution is a wrapping enum (LoopHandlerEvent) that carries the
build_loop::Events as well as NewListener messages.