Adding JSON serde support
Streamed events are designed to be consumed by machine parsers of JSON
(e.g. jq).  Therefore the events that are to be emitted need to be
serializable as JSON. This also includes serializing type names (for
instance) in snake case, and changing a couple of types to be easier to
serialize. Notably, error::BuildError::Exit carries an Option<i32>
instead of a process::ExitError.