build_loop: reorder loop to avoid passing mutable reason around
This should make it a lot easier to follow the code.
We build once at the beginning, and then enter the blocking `select!`,
and after it unblocks we build again and loop around.

This also removes the need of passing `reason` around as a mutable
variable, because it doesn’t have to be carried over to the next loop.