watch: separate file enumeration from self.path_add
Instead of interleaving file enumeration and adding to the watcher, we
create a function which lists paths recursively (`walk_path_topo`),
filter nix paths out of that list and then add it to the watcher in
one go.

The enumeration does not use recursive calls, rather a work queue.

We can add tests that don’t have to go through the watchers now, e.g.
to test the path filtering.