feat(thread.rs): If thread panics, forward the call stack
If a thread paniced before, it would dump its call stack onto stderr,
and then the thread pool would also panic, with its own call stack.

Combined with `human_panic` this would mean that the call stack of the
inner thread would get lost, because it’s not the call stack that
`human_panic` would get to see (it `catch_unwind` to catch the call
stack before it’s printed to stderr by the handler).

The solution is to `catch_unwind` the panic that our inner thread
throws, and to `resume_unwind` from the thread pool, which causes it
to re-panic with the stack collected from the inner thread.

Now we propagate the full call stack.

Previously, the `Drop` implementation of `DeathCertificate` was the
part that ensured that panics would still trigger sending the
DeathCertificate, and then it implicitely (ab)used the fact that the
`join` would return `Err(Any)` on a paniced thread to propagate the
panic. Now we don’t need the death certificate, because the
`catch_unwind` will tell us if there was a panic, so we translate it
to `Cause::Paniced` (or `Cause::Natural` if the thread died a natural
death). This is quite morbid, but also funny.

--

It comes at a cost however, as the `catch/resume_unwind` pair can only
serialize types over thread bounds that are
[`UnwindSafe`](https://doc.rust-lang.org/std/panic/trait.UnwindSafe.html)
Which obviously means mutable references can’t be passed, also most
types using internal mutability, like `JoinHandle`.

Through good fortune, the crossbeam_channel Sender/Receiver is
UnwindSafe (the `std::sync::mpsc` versions are not), so we can still
use channels to communicate over thread boundaries.

--

In `daemon.rs`, we have to do some adjustments to work around the
UnwindSafe restrictions, but with a bit of shuffling it works out and
is not worse than before.

The big problem I faced in `build_instruction_handler` is that it is
the “sink” of the build action channels, and thus blocks on build
messages and spawns more threads. This means we cannot use the pool to
spawn them, because we cannot *both* `join` and wait for channel
messages at the moment.

This could be solved by giving the thread pool a `join_or_spawn`
method, which tries to join the threads while also waiting for
messages to come through a channel, spawning new threads when messages
come in.