fix(treewide): remove serde::Deserialize instances for our types
Earlier we were using the Deserialize instances because our socket
protocol was just the a trivial serde encoder/decoder.

When we switched to varlink, the Serializers and Deserializers
actually come from the varlink Value type, not from our data types, so
we don’t need to implement Deserialize anymore.

Now, we are still using Serialize, for the json streaming feature. We
should probably make that more specific and separate out the event
types for streaming on the command line.