fix(ci): remove additional nix-shell call
We are already calling nix-shell in the next step, even with less
dependencies. So it is not clear at all, what an additional evaluation
of the whole shell (which might take quite some seconds) would give
us.
If we need the artifacts, they are going to be cached when they
are needed. Other than that, no sense in downloading them.