feat(sigterm): daemon exits with 0 on SIGINT or SIGTERM
Make the `lorri daemon` subcommand exit with 0 on SIGINT
or SIGTERM, instead of 130/143.

Other `lorri` subcommands preserve the default SIGINT
and SIGTERM behaviors.

This change is done adds the [ctrlc](https://github.com/Detegr/rust-ctrlc)
crate, by Antti Keränen, as a dependency. This crate provides a simple
and safe interface for handling SIGINT and SIGTERM.

An alternative to the `ctrlc` crate would be to implement the response to the signals
using the the `nix` crate, which is already a lorri dependency.
The downside of this would be adding `unsafe` code to the codebase.

fixes #463